finishing touches for xp activity, migration to add timezone info to db
