+++ charts/agimus/Chart.yaml
@@ -2,5 +2,5 @@ apiVersion: v2
 name: agimus
 description: A helm chart for a discord bot that also runs a mysql db
 type: application
-version: v2.4.4
-appVersion: v2.4.4
+version: v2.5.0
+appVersion: v2.5.0
